requirements.txt
markitdown>=0.0.1a3
openai>=1.0.0
fastapi>=0.109.0
python-multipart>=0.0.6
uvicorn>=0.25.0
streamlit>=1.29.0
requests>=2.31.0
python-dotenv>=1.0.0
